DAO done

+++ src/dao/classes/user.mongo.js
@@ -1,34 +1,39 @@
-import userModel from "./models/user.model.js";
+import userModel from "../models/user.model.js";
 
 export default class User {
-   
+
     get = async () => {
         try {
             let user = await userModel.find()
             return user
-            
+
         } catch (error) {
             console.error(error)
         }
     }
-    
+
     getById = async (userId) => {
         try {
             let user = await userModel.find({ _id: userId })
             return user
-            
+
         } catch (error) {
             console.error(error)
         }
     }
 
-    put = async() => {
-        let updateUser = await userModel.updateOne()
+    put = async (userId, newUserInfo) => {
+        let findUser = await userModel.find({ _id: userId })
+        if (findUser === null) {
+            return res.send("user's ID not found")
+        }
+
+        let updateUser = await findUser.updateOne(newUserInfo)
         return updateUser
     }
 
-    delete = async() => {
-        let deleteUser = await userModel.deleteOne()
+    delete = async (userId) => {
+        let deleteUser = await userModel.deleteOne({ _id: userId })
         return deleteUser
     }
 }
